+++ docs/07-Templates/table.mdx
@@ -8,11 +8,11 @@ draft: true
 
 import { Page } from '@site/src/components/templates/Page/Page';
 import CodeBlock from '@theme/CodeBlock';
-import TablePage from '@site/src/components/templates/TablePage/Table';
+import TablePage from '@site/src/components/templates/TablePage/TablePage';
 import { DataProvider } from '@site/src/components/templates/TablePage/DataProvider';
 import { CodeCollapse } from '@site/src/components/templates/CodeCollapse';
 
-import TablePageRaw from '!!raw-loader!@site/src/components/templates/TablePage/Table';
+import TablePageRaw from '!!raw-loader!@site/src/components/templates/TablePage/TablePage';
 import useColumns from '!!raw-loader!@site/src/components/templates/TablePage/useColumns';
 import types from '!!raw-loader!@site/src/components/templates/TablePage/types';
